+++ tests/Integration/Resources/DSC_SqlDatabaseObjectPermission.config.ps1
@@ -110,6 +110,7 @@ Configuration DSC_SqlDatabaseObjectPermission_Prerequisites_Table1_Config
     {
         SqlScriptQuery 'CreateTable'
         {
+            Id                   = 'CreateTable'
             ServerName           = $Node.ServerName
             InstanceName         = $Node.InstanceName
 
@@ -146,6 +147,7 @@ Configuration DSC_SqlDatabaseObjectPermission_Prerequisites_Procedure1_Config
     {
         SqlScriptQuery 'CreateProcedure1'
         {
+            Id                   = 'CreateProcedure1'
             ServerName           = $Node.ServerName
             InstanceName         = $Node.InstanceName
 
@@ -182,6 +184,7 @@ Configuration DSC_SqlDatabaseObjectPermission_Prerequisites_Procedure2_Config
     {
         SqlScriptQuery 'CreateProcedure2'
         {
+            Id                   = 'CreateProcedure2'
             ServerName           = $Node.ServerName
             InstanceName         = $Node.InstanceName